Return to screen for runner
This started working for some reason.
Part of the reason might be that I deleted my home directory
~/.terminfo/s/st*.
See https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1003257 which gave
me the idea that the st terminfo was the issue.
The issue with `tmux` was that its scrolling behavior wasn't compatible
with my approach for resizing the window using `st`.
In `tmux`, when output was overflowing the buffer, it would scroll back
up to the top and start writing to the top of the buffer rather than
continuing on to the bottom.
For some reason, the tmux implementation would do this only when the
output was very fast, and not when it was given slowly.
I spent a bunch of time tinkering with my patch to get it working with
`tmux`, but then I realized that `screen` worked the way I wanted it to.