Make runner layer not light up keyboard
It was annoying, since it's very obvious when the runner layer is active.
The only issue is that currently, our preprocessor will still create an
exit button which will turn down the keyboard backlight, meaning that if
we have our brightness at 1 to begin with, then enter and exit the run
layer, the ending brightness would be 0.